# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: Global Hospital Management System API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /pharmacy/medicines:
    get:
      operationId: listMedicines
      tags: [pharmacy]
      summary: List medicines inventory
      parameters:
        - in: query
          name: search
          schema:
            type: string
        - in: query
          name: lowStock
          schema:
            type: boolean
        - in: query
          name: expiringSoon
          schema:
            type: boolean
      responses:
        "200":
          description: List of medicines
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Medicine"
    post:
      operationId: createMedicine
      tags: [pharmacy]
      summary: Add medicine to inventory
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/MedicineInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Medicine"
  /pharmacy/medicines/{id}:
    patch:
      operationId: updateMedicine
      tags: [pharmacy]
      summary: Update medicine stock/details
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/MedicineUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Medicine"
    delete:
      operationId: deleteMedicine
      tags: [pharmacy]
      summary: Remove medicine
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted

  # ─────────────────────── PHARMACY BILLS ───────────────────────
  /pharmacy/bills:
    get:
      operationId: listPharmacyBills
      tags: [pharmacy]
      summary: List pharmacy bills
      parameters:
        - in: query
          name: patientId
          schema:
            type: integer
        - in: query
          name: date
          schema:
            type: string
      responses:
        "200":
          description: List of bills
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/PharmacyBill"
    post:
      operationId: createPharmacyBill
      tags: [pharmacy]
      summary: Create a pharmacy bill
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PharmacyBillInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PharmacyBill"
  /pharmacy/bills/{id}:
    get:
      operationId: getPharmacyBill
      tags: [pharmacy]
      summary: Get pharmacy bill
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Bill
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PharmacyBill"
        "404":
          description: Not found

  # ─────────────────────── DASHBOARD ───────────────────────
...
components:
  schemas:
...
    Medicine:
      type: object
      required: [id, name, genericName, category, manufacturer, quantity, unitPrice, expiryDate, batchNumber, createdAt]
      properties:
        id:           { type: integer }
        name:         { type: string }
        genericName:  { type: string }
        category:     { type: string }
        manufacturer: { type: string }
        quantity:     { type: integer }
        unitPrice:    { type: number }
        expiryDate:   { type: string }
        batchNumber:  { type: string }
        minStock:     { type: integer }
        description:  { type: ["string", "null"] }
        createdAt:    { type: string }
    MedicineInput:
      type: object
      required: [name, genericName, category, manufacturer, quantity, unitPrice, expiryDate, batchNumber]
      properties:
        name:         { type: string }
        genericName:  { type: string }
        category:     { type: string }
        manufacturer: { type: string }
        quantity:     { type: integer }
        unitPrice:    { type: number }
        expiryDate:   { type: string }
        batchNumber:  { type: string }
        minStock:     { type: integer }
        description:  { type: string }
    MedicineUpdate:
      type: object
      properties:
        name:         { type: string }
        genericName:  { type: string }
        category:     { type: string }
        manufacturer: { type: string }
        quantity:     { type: integer }
        unitPrice:    { type: number }
        expiryDate:   { type: string }
        batchNumber:  { type: string }
        minStock:     { type: integer }
        description:  { type: string }

    # ─── Pharmacy Bill ───
    PharmacyBillItem:
      type: object
      required: [medicineId, medicineName, quantity, unitPrice, totalPrice]
      properties:
        medicineId:   { type: integer }
        medicineName: { type: string }
        quantity:     { type: integer }
        unitPrice:    { type: number }
        totalPrice:   { type: number }
    PharmacyBill:
      type: object
      required: [id, billNumber, patientId, patientName, totalAmount, discount, netAmount, paymentMode, billDate, createdAt, items]
      properties:
        id:          { type: integer }
        billNumber:  { type: string }
        patientId:   { type: integer }
        patientName: { type: string }
        totalAmount: { type: number }
        discount:    { type: number }
        netAmount:   { type: number }
        paymentMode: { type: string, enum: [cash, card, upi] }
        billDate:    { type: string }
        createdAt:   { type: string }
        items:
          type: array
          items:
            $ref: "#/components/schemas/PharmacyBillItem"
    PharmacyBillInput:
      type: object
      required: [patientId, paymentMode, billDate, items]
      properties:
        patientId:   { type: integer }
        discount:    { type: number }
        paymentMode: { type: string, enum: [cash, card, upi] }
        billDate:    { type: string }
        items:
          type: array
          items:
            type: object
            required: [medicineId, quantity]
            properties:
              medicineId: { type: integer }
              quantity:   { type: integer }

    # ─── Dashboard ───
...