# lib/api-zod/src/generated/types/medicine.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Global Hospital Management System API
 * OpenAPI spec version: 0.1.0
 */

export interface Medicine {
  id: number;
  name: string;
  genericName: string;
  category: string;
  manufacturer: string;
  quantity: number;
  unitPrice: number;
  expiryDate: string;
  batchNumber: string;
  minStock?: number;
  /** @nullable */
  description?: string | null;
  createdAt: string;
}