# lib/api-zod/src/generated/types/pharmacyBill.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Global Hospital Management System API
 * OpenAPI spec version: 0.1.0
 */
import type { PharmacyBillItem } from './pharmacyBillItem';
import type { PharmacyBillPaymentMode } from './pharmacyBillPaymentMode';

export interface PharmacyBill {
  id: number;
  billNumber: string;
  patientId: number;
  patientName: string;
  totalAmount: number;
  discount: number;
  netAmount: number;
  paymentMode: PharmacyBillPaymentMode;
  billDate: string;
  createdAt: string;
  items: PharmacyBillItem[];
}