# lib/api-zod/src/generated/types/pharmacyBillInput.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Global Hospital Management System API
 * OpenAPI spec version: 0.1.0
 */
import type { PharmacyBillInputItemsItem } from './pharmacyBillInputItemsItem';
import type { PharmacyBillInputPaymentMode } from './pharmacyBillInputPaymentMode';

export interface PharmacyBillInput {
  patientId: number;
  discount?: number;
  paymentMode: PharmacyBillInputPaymentMode;
  billDate: string;
  items: PharmacyBillInputItemsItem[];
}